+++ qlearning.h
@@ -13,7 +13,7 @@ void destroy_qtable(struct QTable *qtable);
 void set_state_action_value(struct QTable *qtable, int state, int action, float value);
 float get_state_action_value(struct QTable *qtable, int state, int action);
 float get_max_state_value(struct QTable *qtable, int state);
-float get_best_action(struct QTable *qtable, int state);
+int get_best_action(struct QTable *qtable, int state);
 
 void display_qtable(struct QTable *qtable);
